Propagate flags of secondary to primary
