+++ vulkano/src/command_buffer/auto.rs
@@ -1607,10 +1607,10 @@ impl<P> AutoCommandBufferBuilder<P> {
 
     /// Adds a command that executes a secondary command buffer.
     ///
-    /// # Safety
-    /// The execution state of the secondary command buffer is not checked, and resources are not
-    /// synchronised against concurrent access.
-    pub unsafe fn execute_commands<C>(
+    /// If the `flags` that `command_buffer` was created with are more restrictive than those of
+    /// `self`, then `self` will be restricted to match. E.g. executing a secondary command buffer
+    /// with `Flags::OneTimeSubmit` will set `self`'s flags to `Flags::OneTimeSubmit` also.
+    pub fn execute_commands<C>(
         &mut self,
         command_buffer: C,
     ) -> Result<&mut Self, ExecuteCommandsError>
@@ -1622,14 +1622,20 @@ impl<P> AutoCommandBufferBuilder<P> {
         }
 
         self.check_command_buffer(&command_buffer)?;
+        let secondary_flags = command_buffer.inner().flags();
 
-        {
+        unsafe {
             let mut builder = self.inner.execute_commands();
             builder.add(command_buffer);
             builder.submit()?;
         }
 
+        // Secondary command buffer could leave the primary in any state.
         self.state_cacher.invalidate();
+
+        // If the secondary is non-concurrent or one-time use, that restricts the primary as well.
+        self.flags = std::cmp::min(self.flags, secondary_flags);
+
         Ok(self)
     }
 
@@ -2429,13 +2435,17 @@ impl fmt::Display for AutoCommandBufferBuilderContextError {
 mod tests {
     use crate::buffer::BufferUsage;
     use crate::buffer::CpuAccessibleBuffer;
+    use crate::command_buffer::synced::SyncCommandBufferBuilderError;
     use crate::command_buffer::AutoCommandBufferBuilder;
     use crate::command_buffer::CommandBuffer;
+    use crate::command_buffer::CommandBufferExecError;
+    use crate::command_buffer::ExecuteCommandsError;
     use crate::device::Device;
     use crate::device::DeviceExtensions;
     use crate::device::Features;
+    use crate::instance;
     use crate::sync::GpuFuture;
-    use instance;
+    use std::sync::Arc;
 
     #[test]
     fn copy_buffer_dimensions() {
@@ -2496,5 +2506,63 @@ mod tests {
         let result = destination.read().unwrap();
 
         assert_eq!(*result, [0_u32, 1, 2, 3, 4]);
+    }
+
+    #[test]
+    fn secondary_nonconcurrent_conflict() {
+        let (device, queue) = gfx_dev_and_queue!();
+
+        // Make a secondary CB that doesn't support simultaneous use.
+        let builder =
+            AutoCommandBufferBuilder::secondary_compute(device.clone(), queue.family()).unwrap();
+        let secondary = Arc::new(builder.build().unwrap());
+
+        {
+            let mut builder =
+                AutoCommandBufferBuilder::primary_simultaneous_use(device.clone(), queue.family())
+                    .unwrap();
+
+            // Add the secondary a first time
+            builder.execute_commands(secondary.clone()).unwrap();
+
+            // Recording the same non-concurrent secondary command buffer twice into the same
+            // primary is an error.
+            assert!(matches!(
+                builder.execute_commands(secondary.clone()),
+                Err(ExecuteCommandsError::SyncCommandBufferBuilderError(
+                    SyncCommandBufferBuilderError::ExecError(
+                        CommandBufferExecError::ExclusiveAlreadyInUse
+                    )
+                ))
+            ));
+        }
+
+        {
+            let mut builder =
+                AutoCommandBufferBuilder::primary_simultaneous_use(device.clone(), queue.family())
+                    .unwrap();
+            builder.execute_commands(secondary.clone()).unwrap();
+            let cb1 = builder.build().unwrap();
+
+            let mut builder =
+                AutoCommandBufferBuilder::primary_simultaneous_use(device.clone(), queue.family())
+                    .unwrap();
+
+            // Recording the same non-concurrent secondary command buffer into multiple
+            // primaries is an error.
+            assert!(matches!(
+                builder.execute_commands(secondary.clone()),
+                Err(ExecuteCommandsError::SyncCommandBufferBuilderError(
+                    SyncCommandBufferBuilderError::ExecError(
+                        CommandBufferExecError::ExclusiveAlreadyInUse
+                    )
+                ))
+            ));
+
+            std::mem::drop(cb1);
+
+            // Now that the first cb is dropped, we should be able to record.
+            builder.execute_commands(secondary.clone()).unwrap();
+        }
     }
 }
